add db accessor
